// src/modules/garmentsProduction/reports/embellishment/embellishment-order-summary-report/embellishment-order-summary-report-type.ts
export type EmbellishmentOrderSummaryReportType = {
  EMBELLISHMENT_ORDERNO: string;
  WO_ID: number;
  BUYER_ID: number;
  STYLE_ID: number;
  PO_ID: number;
  OS_BUYER_ID: number;
  OS_STYLE_ID: number;
  OS_PO_ID: number;
  COLOR_ID: number;
  EMBELLISHMENT_TYPE_ID: number;
  SUPPLIER: string;
  WORK_ORDER_NO: string;
  BUYER: string;
  STYLE: string;
  PO_NO: string;
  COLOR: string;
  PO_QTY: number;
  OS_BUYER: string;
  OS_STYLE: string;
  OS_PO_NO: string;
  WORK_ORDER_DATE: string;
  ASKING_DELIVERY_DATE: string;
  RECEIVE_TYPE: string;
  EMBELLISHMENT_TYPE: string;
  EMBELLISHMENT_CATEGORY: string;
  PARTS: string;
  WO_QTY: number;
  RATE: number;
  BUDGET_STATUS: string;
  RCV_QTY: number;
  PRO_QTY: number;
  DELIVERY_QTY: number;
  REJECT_QTY: number;
};